\name{PFoptim-package}
\alias{PFoptim-package}
\alias{PFoptim}
\docType{package}
\title{
  The `PFoptim' package: summary information
}
\description{
  The package provides an implementation of the G-PFSO (Global Particle Filter Stochastic Optimization) algorithm of \insertCite{gerber2020online2;textual}{PFoptim} for finding the global minimizer of a function defined through an expectation. In addition, a function for implementing the SSP resampling algorithm\insertCite{gerber2019negative}{PFoptim} and a function for implementating the Stratified resampling algorithm are also provided.
}

\author{
Mathieu Gerber

Maintainer: Mathieu Gerber <mathieu.gerber@bristol.ac.uk>
}
\references{
  \insertAllCited{}
}

