% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Stratified_Resampler}
\alias{Stratified_Resampler}
\title{Stratified resampling}
\usage{
Stratified_Resampler(U,W)
}
\arguments{
\item{W}{A vector of normalized weights.}

\item{U}{A vector of points in (0,1) such that \code{length(U)=length(W)}.}
}
\value{
A vector of length N with elements in the set \eqn{\{1,...,N\}}, with \code{N=length(U)=length(W)}.
}
\description{
This function implements the stratified resampling algorithm descibed see e.g. in Section 9.6 of \insertCite{chopin2020introduction;textual}{PFoptim}
}
\examples{
N<-100
W<-rbeta(N,0.5,2)
W<-W/sum(W)
J<-Stratified_Resampler(runif(N),W)
}
\references{
\insertAllCited{}
}
\keyword{algorithms}
\keyword{resampling}
